Allow camera settings to change instantly.
Camera settings only changed when a picture was taken, this is still
the current behavior, but this change allows notifications for when
the camere settings are changed in the UI, an may be reflecdet in
the camera, even without taking a picture.

Some code reorganization is also done, so that camera utilities are
separated from general utilities.