image_monitor: Don't process non-image files.
As image_monitor monitors any filesystem event, it would try to open
any new file as if it was an image. By filtering out any file which the
mime type is no 'image/*' no error should be displayed on the terminar,
and the aplpication will not crash for this reason.